save image only when Save Button is clicked
